requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
Django==6.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-nested-admin==4.1.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
Faker==40.1.0
freetype-py==2.5.1
html5lib==1.1
idna==3.11
inflection==0.5.1
lxml==6.0.2
oscrypto==1.3.0
packaging==25.0
pillow==12.0.0
pycairo==1.29.0
pycparser==2.23
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.5.0
python-bidi==0.6.7
python-monkey-business==1.1.0
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.7
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
xhtml2pdf==0.2.17
